kafka: consolidate add partitions to txn handler
