+++ src/v/kafka/CMakeLists.txt
@@ -3,7 +3,6 @@ add_subdirectory(protocol)
 set(handlers_srcs
   server/handlers/api_versions.cc
   server/handlers/metadata.cc
-  server/handlers/add_partitions_to_txn.cc
   server/handlers/txn_offset_commit.cc
   server/handlers/find_coordinator.cc
   server/handlers/describe_configs.cc
